Update Curl.php
Sometimes, e.g. when you want to perform a get request using same name parameters in the query, it is preferred to build the url with the query parameters in the $url argument and not specify a $data argument. In that case, you do not want a '?' put behind your url, as it messes up the query.