Don’t close the cURL handle more than once
When close() was called more than once PHP warnings appeared stating that "XY is not a valid cURL handle resource". This was especially problematic when close() was called manually because it will also *always* be called from the destructor. From no on we’ll first check whether $this->curl is still a resource to avoid that.